requirements.txt
Django==1.2.3
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.3
wsgiref==0.1.2
yolk==0.4.1
django_extensions==0.5
django-object-permissions==1.1
